anchor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
stellar-sdk==13.1.0
python-multipart==0.0.6
jinja2==3.1.2
